requirements.txt
qrcode[pil]>=7.4.2
fpdf2>=2.7.0
Pillow>=10.0.0
tkcalendar>=1.6.1
